React-Vite & Chakra-UI-v2 Setup complete

+++ MERN/Product Store/backend/Controllers/product.controller.js
@@ -54,12 +54,17 @@ export const updateProduct = async (req, res)=>{ // removed: /api/products
 export const deleteProduct = async (req, res)=>{ // removed: /api/products
     const {id} = req.params;
 
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        console.log("Error Updating Product : ", error.message);
+        return res.status(404).json({ success : false, message : "Product Not Found"});
+    }
+    
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "Product Deleted"});
     } catch (error) {
         console.log("Error Deleting Product : ", error.message);
-        res.status(404).json({ success : false, message: "product not found"});
+        res.status(500).json({ success : false, message: "Server Error"});
     }
     
 };
